requirements.txt
Django==3.2
django-colorfield==0.7.2
influxdb_client